# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bakery ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List system users
      parameters:
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a system user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a system user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a system user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---------- EMPLOYEES ----------
...
  /warehouse-items:
    get:
      operationId: listWarehouseItems
      tags: [warehouse]
      summary: List warehouse items
      parameters:
        - name: branchId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WarehouseItem"
    post:
      operationId: createWarehouseItem
      tags: [warehouse]
      summary: Create a warehouse item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseItem"
  /warehouse-items/{id}:
    patch:
      operationId: updateWarehouseItem
      tags: [warehouse]
      summary: Update a warehouse item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseItem"
        "404":
          description: Not found
    delete:
      operationId: deleteWarehouseItem
      tags: [warehouse]
      summary: Delete a warehouse item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /warehouse-movements:
    get:
      operationId: listWarehouseMovements
      tags: [warehouse]
      summary: List warehouse stock movements
      parameters:
        - name: itemId
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WarehouseMovement"
    post:
      operationId: createWarehouseMovement
      tags: [warehouse]
      summary: Record a stock movement (in / out / adjustment); updates item quantity automatically
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseMovementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseMovement"

  # ---------- AUDIT ----------
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [general_manager, branch_manager]
    User:
      type: object
      required: [id, username, fullName, role, branchId]
      properties:
        id: { type: string }
        username: { type: string }
        fullName: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        active: { type: boolean }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [username, password, fullName, role]
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 4 }
        fullName: { type: string, minLength: 1 }
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: ["string", "null"]
        active: { type: boolean }
    UserUpdate:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 4 }
        fullName: { type: string, minLength: 1 }
        role:
          $ref: "#/components/schemas/UserRole"
        branchId:
          type: ["string", "null"]
        active: { type: boolean }
...
    WarehouseItem:
      type: object
      required:
        [
          id,
          name,
          code,
          category,
          unit,
          quantity,
          minQuantity,
          branchId,
          archived,
        ]
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string }
        category: { type: string }
        unit: { type: string }
        quantity: { type: number }
        minQuantity: { type: number }
        supplier: { type: ["string", "null"] }
        purchasePrice: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        branchId: { type: string }
        branchName: { type: ["string", "null"] }
        archived: { type: boolean }
    WarehouseItemInput:
      type: object
      required: [name, code, category, unit, quantity, minQuantity, branchId]
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        quantity: { type: number, minimum: 0 }
        minQuantity: { type: number, minimum: 0 }
        supplier: { type: string }
        purchasePrice: { type: number, minimum: 0 }
        notes: { type: string }
        branchId: { type: string }
    WarehouseItemUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        minQuantity: { type: number, minimum: 0 }
        supplier: { type: string }
        purchasePrice: { type: number, minimum: 0 }
        notes: { type: string }
        archived: { type: boolean }
    WarehouseMovementType:
      type: string
      enum: [in, out, adjustment]
    WarehouseMovement:
      type: object
      required:
        [
          id,
          itemId,
          type,
          quantity,
          quantityBefore,
          quantityAfter,
          date,
        ]
      properties:
        id: { type: string }
        itemId: { type: string }
        itemName: { type: ["string", "null"] }
        type:
          $ref: "#/components/schemas/WarehouseMovementType"
        quantity: { type: number }
        quantityBefore: { type: number }
        quantityAfter: { type: number }
        userName: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        date: { type: string }
    WarehouseMovementInput:
      type: object
      required: [itemId, type, quantity]
      properties:
        itemId: { type: string }
        type:
          $ref: "#/components/schemas/WarehouseMovementType"
        quantity: { type: number, minimum: 0.01 }
        reason: { type: string }
        notes: { type: string }
...